Extract resolvers and mutations for payment
