+++ saleor/graphql/api.py
@@ -9,12 +9,7 @@ from .discount.schema import DiscountMutations, DiscountQueries
 from .order.schema import OrderMutations, OrderQueries
 from .page.schema import PageMutations, PageQueries
 from .product.schema import ProductMutations, ProductQueries
-from .payment.types import Payment, PaymentGatewayEnum
-from .payment.resolvers import (
-    resolve_payments, resolve_payment_client_token)
-from .payment.mutations import (
-    PaymentCapture, PaymentRefund,
-    PaymentVoid)
+from .payment.schema import PaymentMutations, PaymentQueries
 from .shipping.resolvers import resolve_shipping_zones
 from .shipping.types import ShippingZone
 from .shipping.mutations import (
@@ -29,12 +24,7 @@ from .shop.types import Shop
 
 
 class Query(ProductQueries, AccountQueries, CheckoutQueries, DiscountQueries,
-            MenuQueries, OrderQueries, PageQueries):
-    payment = graphene.Field(Payment, id=graphene.Argument(graphene.ID))
-    payment_client_token = graphene.Field(
-        graphene.String, args={'gateway': PaymentGatewayEnum()})
-    payments = PrefetchingConnectionField(
-        Payment, description='List of payments')
+            MenuQueries, OrderQueries, PageQueries, PaymentQueries):
     shop = graphene.Field(Shop, description='Represents a shop resources.')
     shipping_zone = graphene.Field(
         ShippingZone, id=graphene.Argument(graphene.ID, required=True),
@@ -42,16 +32,6 @@ class Query(ProductQueries, AccountQueries, CheckoutQueries, DiscountQueries,
     shipping_zones = PrefetchingConnectionField(
         ShippingZone, description='List of the shop\'s shipping zones.')
     node = graphene.Node.Field()
-
-    def resolve_payment(self, info, id):
-        return graphene.Node.get_node_from_global_id(info, id, Payment)
-
-    def resolve_payment_client_token(self, info, gateway=None):
-        return resolve_payment_client_token(gateway)
-
-    @permission_required('order.manage_orders')
-    def resolve_payments(self, info, query=None, **kwargs):
-        return resolve_payments(info, query)
 
     def resolve_shop(self, info):
         return Shop()
@@ -67,13 +47,9 @@ class Query(ProductQueries, AccountQueries, CheckoutQueries, DiscountQueries,
 
 class Mutations(ProductMutations, AccountMutations, CheckoutMutations,
                 CoreMutations, DiscountMutations, MenuMutations,
-                OrderMutations, PageMutations):
+                OrderMutations, PageMutations, PaymentMutations):
     authorization_key_add = AuthorizationKeyAdd.Field()
     authorization_key_delete = AuthorizationKeyDelete.Field()
-
-    payment_capture = PaymentCapture.Field()
-    payment_refund = PaymentRefund.Field()
-    payment_void = PaymentVoid.Field()
 
     shop_domain_update = ShopDomainUpdate.Field()
     shop_settings_update = ShopSettingsUpdate.Field()
