+++ saleor/graphql/payment/schema.py
@@ -1,0 +1,31 @@
+import graphene
+from graphql_jwt.decorators import permission_required
+
+from ..core.fields import PrefetchingConnectionField
+from .mutations import PaymentCapture, PaymentRefund, PaymentVoid
+from .resolvers import resolve_payment_client_token, resolve_payments
+from .types import Payment, PaymentGatewayEnum
+
+
+class PaymentQueries(graphene.ObjectType):
+    payment = graphene.Field(Payment, id=graphene.Argument(graphene.ID))
+    payment_client_token = graphene.Field(
+        graphene.String, args={'gateway': PaymentGatewayEnum()})
+    payments = PrefetchingConnectionField(
+        Payment, description='List of payments')
+
+    def resolve_payment(self, info, id):
+        return graphene.Node.get_node_from_global_id(info, id, Payment)
+
+    def resolve_payment_client_token(self, info, gateway=None):
+        return resolve_payment_client_token(gateway)
+
+    @permission_required('order.manage_orders')
+    def resolve_payments(self, info, query=None, **kwargs):
+        return resolve_payments(info, query)
+
+
+class PaymentMutations(graphene.ObjectType):
+    payment_capture = PaymentCapture.Field()
+    payment_refund = PaymentRefund.Field()
+    payment_void = PaymentVoid.Field()
